Clean up and document things
